+++ graphdatascience/tests/unit/test_aura_api.py
@@ -871,24 +871,15 @@ def test_parse_tenant_details() -> None:
     )
 
     expected_details = TenantDetails(
-        "42", ds_type="enterprise-ds", regions_per_provider={"gcp": {"eu-west3"}, "aws": {"us-central1", "us-central3"}}
+        "42",
+        cloud_locations={
+            CloudLocation("gcp", "eu-west3"),
+            CloudLocation("aws", "eu-west1"),
+            CloudLocation("aws", "us-central1"),
+            CloudLocation("aws", "us-central3"),
+        },
     )
     assert details == expected_details
-
-
-def test_parse_non_ds_details() -> None:
-    with pytest.raises(
-        RuntimeError,
-        match="Tenant with id `42` cannot create DS instances. Available instances are `{'enterprise-db'}`.",
-    ):
-        TenantDetails.from_json(
-            {
-                "id": "42",
-                "instance_configurations": [
-                    {"type": "enterprise-db", "region": "europe-west1", "cloud_provider": "aws"}
-                ],
-            }
-        )
 
 
 def test_parse_session_info() -> None: