+++ changelog.md
@@ -6,6 +6,8 @@
 
 ## New features
 
+* Add `GdsSessions::available_cloud_locations()` to return available locations for session creation.
+
 
 ## Bug fixes